remove support for user, group and daemonise in the configuration
I posit that most users of a tool like this today will be deploying using tooling that takes care of concerns like the uid, gid and daemonisation - whether that be systemd, docker, kubernetes, or what-have-you. Even in the absence of those tools, my personal view is that those concerns are better handled by a simple wrapping shell script. This allows the configuration and testing surface area of a tool like this to be as small as possible.

This will be a _breaking change_ for current users of this gem, so perhaps it should be part of a v1.x (or v2.x?) release if that's a concern.

I also replaced the word "daemon" in the gemspec with "server" to reflect this attempt at modernisation.